convert from `.vue` to static `.js`
