+++ index.js
@@ -591,7 +591,7 @@ new (class extends Plugin {
             inputs: {
               close,
               layer,
-              message: (new (Vue.extend(require('./components/Changes.vue').default))({
+              message: (new (Vue.extend(require('./components/changes.js').default))({
                 propsData: {
                   commits: commitItems,
                   layer
@@ -1046,7 +1046,7 @@ new (class extends Plugin {
         ...opts,
         id:            "editing-panel",
         title:         opts.title || "plugins.editing.editing_data",
-        internalPanel: new (Vue.extend(require('./components/Editing.vue').default))({
+        internalPanel: new (Vue.extend(require('./components/editing.js').default))({
           state:         this.state,
           resourcesurl:  opts.resourcesUrl || GUI.getResourcesUrl(),
           showcommitbar: undefined === opts.showcommitbar || opts.showcommitbar,
